feat: update component to avoid prop-drilling
