+++ react-class-components/src/pages/detailPage/detailPage.tsx
@@ -1,9 +1,11 @@
-import { useLocation, useNavigate, useOutletContext, useParams, useSearchParams } from 'react-router-dom';
+import { useLocation, useNavigate, useParams, useSearchParams } from 'react-router-dom';
 import DetailCard from '../../components/resultSection/detailCard/detailCard';
 import './detailPage.css';
 import Loader from '../../components/loader/loader';
 import { useEffect } from 'react';
 import { useGetCharacterByNameQuery } from '../../utils/apiSlice';
+import { useSelector } from 'react-redux';
+import { getCurrentPage } from '../../utils/currentPageSlice';
 
 export default function DetailPage() {
     const { name } = useParams();
@@ -15,19 +17,10 @@ export default function DetailPage() {
         error,
     } = useGetCharacterByNameQuery({ name: name || '' }, { skip: !name });
     const navigate = useNavigate();
-    const page: number = useOutletContext();
+    const page: number = useSelector(getCurrentPage);
     const [, setSearchParams] = useSearchParams();
     const location = useLocation();
     const newURL = `${location.search}${location.hash}`;
-
-    // useEffect(() => {
-    //     if (isSuccess && character && character.length > 0) {
-    //         dispatch(setSelectedItem(character[0]));
-    //     }
-    //     return () => {
-    //         dispatch(clearSelectedItem());
-    //     };
-    // }, [isSuccess, character, dispatch]);
 
     useEffect(() => setSearchParams({ page: `${page}` }), [page, name]);
